ページレイアウトとスタイルの移行: _document.tsx, layout.tsx, globals.css
