--- src/app/layout.tsx
@@ -1,16 +1,0 @@
-"use client";
-
-import "./globals.css";
-import { ReactNode } from "react";
-
-interface RootLayoutProps {
-    children: ReactNode;
-}
-
-export default function RootLayout({ children }: RootLayoutProps) {
-    return (
-        <html lang="ja">
-            <body className="antialiased">{children}</body>
-        </html>
-    );
-}
